requirements.txt
asgiref==3.7.2
black==23.7.0
click==8.1.4
crispy-bootstrap5==0.7
Django==4.2.2
django-crispy-forms==2.0
django-phone-field==1.8.1
django-widget-tweaks==1.4.12
install==1.3.5
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==10.0.0
platformdirs==3.8.1
python-barcode==0.15.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.3
